Edited names on driver hub and moved testings programs into testing folder
