+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MainAutonomous.java
@@ -1,9 +1,12 @@
 package org.firstinspires.ftc.teamcode;
 
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import java.util.Set;
 
+@Autonomous(name="Main Autonomous", group=" Main", preselectTeleOp="Main Teleop")
 public class MainAutonomous extends LinearOpMode {
     Robot robot = new Robot(this);
