+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drivebase/MecanumDrive.java
@@ -1,15 +1,15 @@
 package org.firstinspires.ftc.teamcode.drivebase;
 
-import static org.firstinspires.ftc.teamcode.Utility.applyDeadzone;
-import static org.firstinspires.ftc.teamcode.Utility.normalizePowers;
-
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
+import static org.firstinspires.ftc.teamcode.Utility.applyDeadzone;
+import static org.firstinspires.ftc.teamcode.Utility.normalizePowers;
+
 import org.firstinspires.ftc.teamcode.Robot;
 
-@TeleOp(name="Field-Centric Drive", group="Drivebase")
-public class TeleopFieldCentricDrive extends LinearOpMode {
+@TeleOp(name="Mecanum Drive", group="DriveBase")
+public class MecanumDrive extends LinearOpMode {
     private static final double DEADZONE_THRESHOLD = 0.1;
 
     Robot robot = new Robot(this);
@@ -23,12 +23,8 @@ public class TeleopFieldCentricDrive extends LinearOpMode {
         robot.imu.resetYaw();
 
         while (opModeIsActive()) {
-            if (gamepad1.options) {
-                robot.imu.resetYaw();
-            }
-
             readSensors();
-            FieldCentricDrive();
+            mecanumDrive(straight, turn, strafe);
             updateTelemetryData();
         }
     }
@@ -40,19 +36,12 @@ public class TeleopFieldCentricDrive extends LinearOpMode {
         heading = robot.getHeading(); // in radians
     }
 
-    /**
-     * @see <a href="https://gm0.org/en/latest/docs/software/tutorials/mecanum-drive.html">gm0 guide</a>
-     */
-    private void FieldCentricDrive() {
-        double rotY = strafe * Math.sin(-heading) + straight * Math.cos(-heading);
-        double rotX = strafe * Math.cos(-heading) - straight * Math.sin(-heading);
-
-
+    private void mecanumDrive(double straight, double turn, double strafe) {
         // calculate powers
-        powers[0] = rotY + rotX + turn; // front left power
-        powers[1] = rotY - rotX + turn; // back left power
-        powers[2] = rotY - rotX - turn; // front right power
-        powers[3] = rotY + rotX - turn; // back right power
+        powers[0] = straight + strafe + turn; // front left power
+        powers[1] = straight - strafe + turn; // back left power
+        powers[2] = straight - strafe - turn; // front right power
+        powers[3] = straight + strafe - turn; // back right power
 
         // powers array is updated inside this method
         normalizePowers(powers);
