+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testing/ServoPositionLocator.java
@@ -1,9 +1,12 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.testing;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.Servo;
+
+import org.firstinspires.ftc.teamcode.Utility;
+
 import java.text.DecimalFormat;
 
 @TeleOp(name="Servo Position Locator",group="Testing")
